+++ src/main/java/com/salesforce/dataloader/client/transport/HttpTransportInterface.java
@@ -24,14 +24,14 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.dataloader.client;
+package com.salesforce.dataloader.client.transport;
 
 import java.io.OutputStream;
-import java.net.HttpURLConnection;
 import java.io.InputStream;
 import java.io.IOException;
 import java.util.HashMap;
-import java.util.Map;
+
+import org.apache.http.HttpResponse;
 
 import com.salesforce.dataloader.exception.HttpClientTransportException;
 import com.sforce.async.AsyncApiException;
@@ -44,19 +44,21 @@ import com.sforce.ws.transport.Transport;
  * @version 1.0
  * @since 1.0  Nov 30, 2005
  */
-public interface HttpTransportInterfacev1 extends Transport {
+public interface HttpTransportInterface extends Transport {
 	enum SupportedHttpMethodType {
 		PUT,
 		POST,
 		PATCH,
-		DELETE
+		DELETE,
+		GET
 	}
+    int PROXY_AUTHENTICATION_REQUIRED = 407;
     OutputStream connect(String endpoint, HashMap<String, String> httpHeaders, boolean enableCompression,
-    		HttpTransportInterfacev1.SupportedHttpMethodType httpMethod) throws IOException;
+    		HttpTransportInterface.SupportedHttpMethodType httpMethod) throws IOException;
 
     void connect(String endpoint, HashMap<String, String> httpHeaders, boolean enableCompression,
-    		HttpTransportInterfacev1.SupportedHttpMethodType httpMethod, InputStream contentInputStream, String contentEncoding) throws IOException;
+    		HttpTransportInterface.SupportedHttpMethodType httpMethod, InputStream contentInputStream, String contentEncoding) throws IOException;
     
-    HttpURLConnection openHttpGetConnection(String urlStr, Map<String, String> headers) throws IOException;
-    InputStream httpGet(HttpURLConnection connection, String urlStr) throws IOException, AsyncApiException, HttpClientTransportException;
+    InputStream httpGet(String urlStr) throws IOException, AsyncApiException, HttpClientTransportException;
+    HttpResponse getHttpResponse();
 }
