+++ src/main/java/com/salesforce/dataloader/client/transport/TransportFactoryImpl.java
@@ -23,43 +23,19 @@
  * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  * POSSIBILITY OF SUCH DAMAGE.
  */
-package com.salesforce.dataloader.client;
+package com.salesforce.dataloader.client.transport;
 
-import com.salesforce.dataloader.config.AppConfig;
-import org.apache.http.message.BasicNameValuePair;
+import com.sforce.ws.transport.Transport;
+import com.sforce.ws.transport.TransportFactory;
 
-import java.util.function.Function;
+public class TransportFactoryImpl implements TransportFactory {
 
-/**
- * Created by rmazzeo on 12/9/15.
- */
-public class SimplePostFactory {
-
-    private static Function<Criteria, SimplePostInterface> constructor = c -> new SimplePostImpl(c.appConfig, c.endpoint, c.pairs);;
-
-    public static class Criteria{
-        public AppConfig appConfig;
-        public String endpoint;
-        public BasicNameValuePair[] pairs;
+    public TransportFactoryImpl() {
     }
 
-    public static Function<Criteria, SimplePostInterface> getConstructor()
-    {
-        return constructor;
+    @Override
+    public Transport createTransport() {
+        return HttpTransportImpl.getInstance();
     }
 
-    public static void setConstructor(Function<Criteria, SimplePostInterface> constructor)
-    {
-        SimplePostFactory.constructor = constructor;
-    }
-
-    public static SimplePostInterface getInstance(AppConfig appConfig, String endpoint, BasicNameValuePair... pairs){
-        Criteria criteria = new Criteria();
-        criteria.appConfig = appConfig;
-        criteria.endpoint = endpoint;
-        criteria.pairs = pairs;
-
-        return constructor.apply(criteria);
-
-    }
 }
