+++ src/main/java/com/salesforce/dataloader/util/OAuthBrowserLoginRunner.java
@@ -48,8 +48,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.FieldNamingPolicy;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.salesforce.dataloader.client.SimplePostInterface;
-import com.salesforce.dataloader.client.SimplePostFactory;
+import com.salesforce.dataloader.client.transport.SimplePostFactory;
+import com.salesforce.dataloader.client.transport.SimplePostInterface;
 import com.salesforce.dataloader.config.AppConfig;
 import com.salesforce.dataloader.config.Messages;
 import com.salesforce.dataloader.exception.OAuthBrowserLoginRunnerException;
