+++ src/test/java/com/salesforce/dataloader/oauth/OAuthSecretFlowUtilTests.java
@@ -29,8 +29,8 @@ import com.google.gson.FieldNamingPolicy;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.salesforce.dataloader.ConfigTestBase;
-import com.salesforce.dataloader.client.SimplePostInterface;
-import com.salesforce.dataloader.client.SimplePostFactory;
+import com.salesforce.dataloader.client.transport.SimplePostFactory;
+import com.salesforce.dataloader.client.transport.SimplePostInterface;
 import com.salesforce.dataloader.config.AppConfig;
 import com.salesforce.dataloader.exception.ParameterLoadException;
 import com.salesforce.dataloader.model.OAuthToken;
