+++ src/test/java/com/salesforce/dataloader/process/ProcessTestBase.java
@@ -45,8 +45,8 @@ import org.junit.Before;
 import com.salesforce.dataloader.*;
 import com.salesforce.dataloader.action.OperationInfo;
 import com.salesforce.dataloader.action.progress.ILoaderProgress;
-import com.salesforce.dataloader.client.HttpTransportImpl;
 import com.salesforce.dataloader.client.PartnerClient;
+import com.salesforce.dataloader.client.transport.HttpTransportImpl;
 import com.salesforce.dataloader.config.AppConfig;
 import com.salesforce.dataloader.controller.Controller;
 import com.salesforce.dataloader.dao.DataAccessObjectFactory;
